refactor: centralize type mapping in Scala (DataType) and Rust (conversion module)
