+++ core/src/main/scala/com/github/chitralverma/polars/api/types/Schema.scala
@@ -4,7 +4,6 @@ import scala.jdk.CollectionConverters._
 import scala.util.Try
 
 import com.fasterxml.jackson.databind.JsonNode
-import com.fasterxml.jackson.databind.node.JsonNodeType
 import com.github.chitralverma.polars.jsonMapper
 
 case class Field(name: String, dataType: DataType) {
@@ -39,75 +38,6 @@ class Schema {
 
   override def toString: String = treeString
 
-  private def toField(field: (String, JsonNode, JsonNodeType)): Field = field match {
-    // For Basic Types
-    case (name, node, _ @JsonNodeType.STRING) =>
-      Field(name, DataType.fromBasicType(node.textValue()))
-
-    // For Time Type
-    case (name, node, _ @JsonNodeType.OBJECT)
-        if node.hasNonNull("Time") || node.hasNonNull("Time32") || node.hasNonNull("Time64") =>
-      Seq(node.get("Time"), node.get("Time32"), node.get("Time64"))
-        .map(Option(_))
-        .collectFirst { case Some(v) => v } match {
-        case Some(timeUnit) => Field(name, TimeType(timeUnit.textValue()))
-
-        case None =>
-          throw new IllegalArgumentException("Invalid time cannot be parsed.")
-      }
-
-    // For Duration Type
-    case (name, node, _ @JsonNodeType.OBJECT) if node.hasNonNull("Duration") =>
-      val timeUnit = node.get("Duration")
-      Field(name, Duration(timeUnit.textValue()))
-
-    // For DateTime Type
-    case (name, node, _ @JsonNodeType.OBJECT) if node.hasNonNull("Timestamp") =>
-      node.get("Timestamp").elements().asScala.map(_.asText(null)).toSeq match {
-        case Seq(tu, tz) =>
-          Field(name, DateTimeType(tu, tz))
-        case _ =>
-          Field(name, DateTimeType(null, null))
-      }
-
-    // For (Nested) List Type
-    case (name, node, _ @JsonNodeType.OBJECT)
-        if node.hasNonNull("List") || node.hasNonNull("LargeList") =>
-      Seq(node.get("List"), node.get("LargeList"))
-        .map(Option(_))
-        .collectFirst { case Some(v) => v } match {
-        case Some(listNode) =>
-          val listNodeType = listNode.get("dtype")
-          Field(name, ListType(toField((name, listNodeType, listNodeType.getNodeType)).dataType))
-
-        case None =>
-          throw new IllegalArgumentException("Invalid list cannot be parsed as a JSON.")
-      }
-
-    // For (Nested) Struct Type
-    case (name, node, _ @JsonNodeType.OBJECT) if node.has("Struct") =>
-      val structNode = node.get("Struct")
-      val structFields = structNode.iterator().asScala
-      val sf = structFields.map {
-        case node: JsonNode if node.hasNonNull("name") && node.hasNonNull("dtype") =>
-          val structFieldName: String = node.get("name").textValue()
-          val structFieldType: JsonNode = node.get("dtype")
-
-          Field(
-            structFieldName,
-            toField(name, structFieldType, structFieldType.getNodeType).dataType
-          )
-
-        case _ =>
-          throw new IllegalArgumentException("Invalid struct cannot be parsed as a JSON.")
-      }.toArray
-
-      Field(name, StructType(sf))
-
-    case _ =>
-      throw new IllegalArgumentException("Invalid field cannot be parsed as a JSON.")
-  }
-
   private def setFields(fields: Array[Field]): Schema = {
     fields match {
       case f if f == null || f.isEmpty =>
@@ -129,7 +59,7 @@ class Schema {
         val fields = node.get("fields").elements().asScala.toList
         _fields = fields
           .map(f =>
-            toField(f.get("name").textValue(), f.get("dtype"), f.get("dtype").getNodeType)
+            Field(f.get("name").textValue(), DataType.fromJson(f.get("dtype")))
           )
           .toArray
