+++ native/src/internal_jni/io/scan/ipc.rs
@@ -8,7 +8,8 @@ use polars::io::{HiveOptions, RowIndex};
 use polars::prelude::*;
 
 use crate::internal_jni::io::{get_file_path, parse_json_to_options};
-use crate::internal_jni::utils::{JavaArrayToVec, to_ptr};
+use crate::internal_jni::utils::{to_ptr};
+use crate::internal_jni::conversion::JavaArrayToVec;
 use crate::utils::error::ResultExt;
 
 #[jni_fn("com.github.chitralverma.polars.internal.jni.io.scan$")]
