+++ native/src/internal_jni/series.rs
@@ -11,8 +11,9 @@ use jni_fn::jni_fn;
 use polars::prelude::*;
 
 use crate::internal_jni::utils::{
-    JavaArrayToVec, from_ptr, j_object_to_string, j_string_to_string, to_ptr,
+    from_ptr, j_object_to_string, j_string_to_string, to_ptr
 };
+use crate::internal_jni::conversion::JavaArrayToVec;
 use crate::utils::error::ResultExt;
 
 #[jni_fn("com.github.chitralverma.polars.internal.jni.series$")]
